context-1-data-gen/pyproject.toml
[project]
name = "agentic-search-data-gen"
version = "0.1.0"
description = "Synthetic search task generation across multiple domains using agentic LLM pipelines"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.75.0",
    "chromadb>=1.4.0",
    "openai>=2.15.0",
    "tiktoken>=0.12.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
rerank = ["baseten-performance-client>=0.1.1"]
patents = [
    "python-docx>=1.2.0",
    "pandas>=2.3.3",
    "httpx[http2]>=0.28.1",
    "datalab-python-sdk>=0.2.1",
]
indexing = ["fastembed>=0.7.4"]
epstein = ["gdown>=5.0.0"]
sec-index = ["edgartools>=3.0.0", "fastembed>=0.7.4"]
all = ["agentic-search-data-gen[rerank,patents,indexing,epstein,sec-index]"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
